chronos-backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
groq
python-dotenv
instructor
pydantic
pytz
datetime